+++ src/pages/ingame/ingame-server.ts
@@ -1,11 +1,12 @@
 // src/pages/ingame/ingame.ts
 import "../../style.css";
 import "./ingame-ui.ts";
-import { sendMsg, socket, joinRoom, getRoomInfo } from "./A13C-chat.ts";
+import { sendMsg, socket, joinRoom } from "./A13C-chat.ts";
 import type { RoomMembers, JoinRoomParams } from "./A13C-chat.ts";
 import "./chat.ts";
+import { setupSubmitCardClick } from "./ingame-ui.ts";
 
-function loadCurrentRoom() {
+export function loadCurrentRoom() {
   const saved = sessionStorage.getItem("A13C_CURRENT_ROOM");
   if (!saved) return null;
   try {
@@ -63,10 +64,10 @@ document.addEventListener("DOMContentLoaded", () => {
   updateOpponentNicknames();
 });
 
-interface TempCard {
-  card: number;
-  returnRound: number;
-}
+// interface TempCard {
+//   card: number;
+//   returnRound: number;
+// }
 
 interface Step1Payload {
   action: "step1";
@@ -75,28 +76,11 @@ interface Step1Payload {
   card2: number;
 }
 
-let currentRound = 1;
-let selectedCardNumbers: number[] = [];
-let activeCardId: "left" | "right" | null = null;
-let mySubmittedCard: number | null = null;
-let tempStorage: TempCard[] = [];
-const submittedCards: { [playerId: string]: number } = {};
-
 const selectedLeft = document.getElementById("selected-left") as HTMLDivElement;
 const selectedRight = document.getElementById(
   "selected-right"
 ) as HTMLDivElement;
-const resetBtn = document.querySelector(
-  "#submitbutton button:nth-child(1)"
-) as HTMLButtonElement;
-const submitBtn = document.querySelector(
-  "#submitbutton button:nth-child(2)"
-) as HTMLButtonElement;
-const myCardContainer = document.getElementById("my-cards") as HTMLDivElement;
-const scoreBoard = document.getElementById("final-card-area") as HTMLDivElement;
-const tempStorageArea = document.getElementById(
-  "temp-card-area"
-) as HTMLDivElement;
+
 
 const isHost = localStorage
   .getItem("A13C_CREATE_ROOM_INFO")
@@ -115,102 +99,9 @@ overlay.innerHTML = isHost
   : `<p class="pointer-events-none">방장이 게임을 시작할 때까지 기다려주세요</p>`;
 document.body.appendChild(overlay);
 
-function updateHandCardAvailability(): void {
-  const cards = Array.from(
-    myCardContainer.querySelectorAll<HTMLImageElement>("img[data-card]")
-  );
-  const used = new Set<number>();
-  [scoreBoard, tempStorageArea].forEach((board) => {
-    board
-      .querySelectorAll<HTMLImageElement>("img[data-card]")
-      .forEach((img) => {
-        const m = img.src.match(/card-(\d+)\.webp/);
-        if (m) used.add(Number(m[1]));
-      });
-  });
-  cards.forEach((card) => {
-    const num = Number(card.getAttribute("data-card"));
-    if (used.has(num)) {
-      card.classList.add("opacity-50");
-      card.style.pointerEvents = "none";
-    } else {
-      card.classList.remove("opacity-50");
-      card.style.pointerEvents = "";
-    }
-  });
-}
 
-function renderMyCards(): void {
-  myCardContainer.innerHTML = "";
-  for (let i = 1; i <= 8; i++) {
-    if (selectedCardNumbers.includes(i)) continue;
-    const card = document.createElement("img");
-    card.src = `/imges/card-${i}.webp`;
-    card.className =
-      "w-[153px] h-[214px] cursor-pointer transition-transform duration-200 ease-in-out hover:scale-110";
-    card.setAttribute("data-card", String(i));
-    card.addEventListener("click", () => selectCard(i, card));
-    myCardContainer.appendChild(card);
-  }
-  updateHandCardAvailability();
-}
 
-function selectCard(
-  num: number,
-  cardEl: HTMLImageElement,
-  force = false
-): void {
-  if (cardEl.classList.contains("opacity-50")) return;
-  if (selectedCardNumbers.length >= 2) return;
-  selectedCardNumbers.push(num);
-  cardEl.remove();
-  const target =
-    selectedCardNumbers.length === 1 ? selectedLeft : selectedRight;
-  target.style.backgroundImage = `url("/imges/card-${num}.webp")`;
-  target.setAttribute("data-card-src", `/imges/card-${num}.webp`);
-}
-
-function setupSubmitCardClick(): void {
-  selectedLeft.addEventListener("click", () => {
-    selectedLeft.classList.add("border-4", "border-yellow-300");
-    selectedRight.classList.remove("border-4", "border-yellow-300");
-    activeCardId = "left";
-  });
-  selectedRight.addEventListener("click", () => {
-    selectedRight.classList.add("border-4", "border-yellow-300");
-    selectedLeft.classList.remove("border-4", "border-yellow-300");
-    activeCardId = "right";
-  });
-}
-
-function flyCard(fromEl: HTMLElement, toEl: HTMLElement, src: string): void {
-  const fromRect = fromEl.getBoundingClientRect();
-  const toRect = toEl.getBoundingClientRect();
-  const card = document.createElement("img");
-  card.src = src;
-  const m = src.match(/card-(\d+)\.webp/);
-  if (m) card.setAttribute("data-card", m[1]);
-  card.className = "card-fly";
-  card.style.left = `${fromRect.left}px`;
-  card.style.top = `${fromRect.top}px`;
-  document.body.appendChild(card);
-  requestAnimationFrame(() => {
-    card.style.transform = `translate(${toRect.left - fromRect.left}px, ${toRect.top - fromRect.top}px)`;
-  });
-
-  setTimeout(() => {
-    card.remove();
-    const finalCard = document.createElement("img");
-    finalCard.src = src;
-    finalCard.className = "w-[80px] h-[110px]";
-    if (m) finalCard.setAttribute("data-card", m[1]);
-    toEl.innerHTML = "";
-    toEl.appendChild(finalCard);
-    updateHandCardAvailability();
-  }, 500);
-}
-
-function sendStep1Cards(): void {
+export function sendStep1Cards(): void {
   const card1 = Number(
     selectedLeft.getAttribute("data-card-src")?.match(/\d+/)?.[0]
   );
@@ -226,47 +117,6 @@ function sendStep1Cards(): void {
   socket.emit("message", message);
   console.log("emit step1 message:", message);
 }
-
-resetBtn.addEventListener("click", () => {
-  selectedCardNumbers = [];
-  activeCardId = null;
-  selectedLeft.style.backgroundImage = `url("/imges/card-back.webp")`;
-  selectedRight.style.backgroundImage = `url("/imges/card-back.webp")`;
-  selectedLeft.removeAttribute("data-card-src");
-  selectedRight.removeAttribute("data-card-src");
-  selectedLeft.classList.remove("border-4", "border-yellow-300");
-  selectedRight.classList.remove("border-4", "border-yellow-300");
-  scoreBoard.innerHTML = "";
-  tempStorageArea.innerHTML = "";
-  renderMyCards();
-});
-
-submitBtn.addEventListener("click", () => {
-  const keepEl = activeCardId === "left" ? selectedLeft : selectedRight;
-  const removeEl = activeCardId === "left" ? selectedRight : selectedLeft;
-  const keepSrc = keepEl.getAttribute("data-card-src");
-  const removeSrc = removeEl.getAttribute("data-card-src");
-  if (!activeCardId || !keepSrc || keepSrc.includes("card-back")) {
-    alert("카드를 선택해주세요");
-    return;
-  }
-  flyCard(keepEl, scoreBoard, keepSrc);
-  const m = keepSrc.match(/card-(\d+)\.webp/);
-  if (m) {
-    mySubmittedCard = Number(m[1]);
-    submittedCards["me"] = mySubmittedCard;
-  }
-  if (removeSrc) {
-    const m2 = removeSrc.match(/card-(\d+)\.webp/);
-    if (m2) {
-      const num = Number(m2[1]);
-      tempStorage.push({ card: num, returnRound: currentRound + 2 });
-      flyCard(removeEl, tempStorageArea, removeSrc);
-    }
-  }
-  selectedCardNumbers = [];
-  activeCardId = null;
-});
 
 export function removeOverlay(): boolean {
   const overlay = document.getElementById("game-overlay");
@@ -369,3 +219,6 @@ socket.on("members", (members: RoomMembers) => {
     }
   }, 300);
 });
+
+// 초기 카드 렌더링
+sendStep1Cards();
